cs493/hw3-chanall/requirements.txt
Flask==2.2.2
google-cloud-datastore==1.7.3